main: gracefully cease HTTP serving on shutdown. (#34)
When the controller context is cancelled and the worker goroutines have
ceased serving, the HTTP serving interface should attempt a graceful
shutdown, but not before.  This pattern sets the stage for probable
application of the same to the SSH interface.

There are a few potential avenues for additional cleanups around
monitoring interface setup and context plumbing that should be explored
independently.  I wanted to keep the code delta as small as possible
in case I made any mistakes.