SSH: allow RO keys
Make keys ro by annotating them in the config.toml file and check if the
operation performed is an RO one (currently only list).

This helps for writing a prober that exports the diff status from
looking at this output and knowing the latest git commit. This probably
means *we* don't have to export the git hash, but only list it from the
API.

Probably warrants a small other cmd/ that actually does these checks and
sets the 'diff detected' metrics that will then picked up.

Signed-off-by: Miek Gieben <miek@science.ru.nl>