There is a bug with paths longer than 80 characters. When running `git pull --stat` for changes with these long paths. You get a truncated path:
```
.../deeply/nested/directory/hierarchy/this.service | 6 +++---
```

This will not match the set `dir` in the gitopper configuration.

Unfortunately you cannot set the maximum width in `git pull --stat` but possible with `git diff --stat`. We have to separate the steps when checking for changes. That is, `git fetch`, `git diff`, then `git merge`. The default kernel-imposed limit in Linux is 4096 so just use that too.

It does `git diff` now so we can also set `--name-only` to get a clean output for string matching. We can skip matching the path inside the `git pull --stat` output format.